requirements.txt
python-telegram-bot==21.4
python-dotenv==1.0.1
google-generativeai==0.7.2
Pillow==10.4.0
